BackEnd/requirements.txt
# AWS SDK for Python (required for S3, Transcribe, Lambda)
boto3>=1.34.0
botocore>=1.34.0

# Environment variable management
python-dotenv>=1.0.0

# HTTP requests (for downloading transcripts)
requests>=2.31.0

# Google Gemini API (for audio processing)
google-generativeai>=0.3.2

# Type hints and validation
pydantic>=2.8.0

# Utilities
python-dateutil>=2.8.2
